chore: address review comments

+++ crates/miden-protocol/src/errors/mod.rs
@@ -432,7 +432,7 @@ pub enum AccountDeltaError {
     )]
     TooManyVaultAssetDeltas {
         delta_op: AssetDeltaOperation,
-        num_ops: usize,
+        num_ops: u16,
     },
     #[error(
         "account update of type `{left_update_type}` cannot be merged with account update of type `{right_update_type}`"
